More flexible structure

+++ memeocr.py
@@ -9,12 +9,32 @@ class MemeOCR:
         self._tmp_image_fname = '/tmp/memeocr.jpg'
         self._tmp_txt_base = '/tmp/memeocr'
         self._tmp_txt_fname = self._tmp_txt_base + '.txt'
+        self._template_image = None
+        self._keep_tmp_files = False
 
-    def _thresh_words(self, fname):
+    def set_template(self, fname):
+        self._template_image = self._read_image(fname)
+
+    def recognize(self, fname):
+        txt = None
+        img = self._read_image(fname)
+        self._thresh_words(img, self._template_image)
+        self._exec_tesseract()
+        txt = self._read_txt()
+        self._delete_tmp_files()
+        return txt
+
+    def _read_image(self, fname):
         try:
             img = cv2.imread(fname)
         except IOError:
-            return False
+            img = None
+        return img
+
+    def _thresh_words(self, img, template):
+        if img is None:
+            return
+
         for i in range(img.shape[0]):
             for j in range(img.shape[1]):
                 if all([elem >= self._white_thresh for elem in img[i][j]]):
@@ -23,12 +43,10 @@ class MemeOCR:
                     img[i][j] = (0, 0, 0)
 
         cv2.imwrite(self._tmp_image_fname, img)
-        return True
 
     def _exec_tesseract(self):
         cmd = 'env TESSDATA_PREFIX=./ tesseract -l joh %s %s > /dev/null' % (self._tmp_image_fname, self._tmp_txt_base)
-        ret = os.system(cmd)
-        return ret == 0
+        os.system(cmd)
 
     def _read_txt(self):
         try:
@@ -42,19 +60,10 @@ class MemeOCR:
         return lines
 
     def _delete_tmp_files(self):
+        if self._keep_tmp_files:
+            return
         if os.path.exists(self._tmp_image_fname):
             os.remove(self._tmp_image_fname)
         if os.path.exists(self._tmp_txt_fname):
             os.remove(self._tmp_txt_fname)
-
-    def recognize(self, fname):
-        if not self._thresh_words(fname):
-            self._delete_tmp_files()
-            return None
-        if not self._exec_tesseract():
-            self._delete_tmp_files()
-            return None
-        txt = self._read_txt()
-        self._delete_tmp_files()
-        return txt
